requirements.txt
python-telegram-bot>=20.0
httpx>=0.27.0
rapidfuzz>=3.0.0
